buyit/requirements.txt
asgiref @ file:///C:/ci_311/asgiref_1676448296195/work
Django @ file:///C:/ci_311/django_1676469589007/work
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg2==2.9.7
python-dotenv==1.0.0
pytz==2023.3
sqlparse @ file:///C:/b/abs_88361ub_qu/croot/sqlparse_1690904577514/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
